Worn armor has a 25% weight reduction
Drawn from a FIQHack feature that does a similar thing. The idea here is
that when you are wearing armor, various muscles not usually used for
carrying gear are doing some of the lifting work. (Think of wearing a
50-pound suit of armor and moving around, versus carrying 50 pounds of
concrete in one's arms or a knapsack.)

However, in FIQHack, armor weighs 50% as much when worn, but armor
weights were increased to compensate. I ignored that because I really
don't want to try and throw armor weight rebalance into this. Thus, the
base weights are unchanged but the percentage reduction is less.