Fix: getobj didn't allow downplayed '-' to be selected
When a callback returns GETOBJ_DOWNPLAY, the item is intended to be
selectable so that it can be returned from getobj. This didn't work
correctly for selecting '-' (hands/self) - it would only be selectable
if the callback returned GETOBJ_SUGGEST. The same goes for
GETOBJ_EXCLUDE_INACCESS and GETOBJ_EXCLUDE_SELECTABLE, which are
supposed to allow the object (or lack thereof in this case) to be
selected.

In those three cases, this adjusts getobj's behavior so that it becomes
selectable.