Fix: thiefstones not allowing some items to stack in stash
This was a result of self-proclaimed hacks that set no_charge on the
item to get around having items added to a shop bill that shouldn't get
put on the bill (e.g. a player-owned thiefstone teleporting a shop-owned
item out of a shop should bill for only the shop-owned item, whereas the
reverse scenario should bill for only the thiefstone).

Since thiefstone_teleport can choose not to call addtobill in the first
place, it was fairly easy to get rid of those hacks by just adding a
third argument to thiefstone_teleport.

In the process I realized that the ability of players to have a
thiefstone keyed to wherever the player wants contradicts earlier
assumptions made that a thiefstone would never be keyed into a shop. So
I fixed a known issue with that scenario, in which if you teleport
shop-owned stuff to the keyed location while in the same shop, the
shopkeeper charges you for it, even though they still have the item. All
other cases that I know of involving teleportation into shops all result
in the shopkeeper laying claim to the item, which makes sense.