Refer to create_critters' gendered groups properly
Not really a bug, but it would produce a message e.g. "Dwarf leaders
appear from nowhere!" even if the group was all-male or all-female. This
tracks the gender count, and will use the appropriate gender in that
message if the group is all one gender.

Does not affect cases where create_critters is called multiple times
with NULL as the monster pointer and happens to create the same monster
species - that will still just be "Monsters appear from nowhere!"