Fix: shop generation segfault trying to find shopkeeper name
This happened because hash1() returns an unsigned int, but the local
variable for it was an int. When this conversion took it negative, it
tried to use a negative index into the names array and crashed.

I also addressed several other potential misuses of hash1() in the same
vein: where possible, just converting the relevant variables to unsigned
is sufficient. But in a couple cases, a hash value is passed to a
function which takes a signed int that uses a negative value to mean it
wants a random value. In these cases it's necessary to make sure the int
value is positive, so I added a function int_hash1() that basically does
that, and used it in those cases.