Move sea monsters back to ; and restore zruties
This is a significant move, and a rather large reversal on my part. This
was one of xNetHack's distinctive features. Unfortunately, I now think
it's distinctive in a bad way. For all the time this has been present in
xNetHack, player opinions have largely been negative, varying from those
that are just annoyed at the disruption (why doesn't blessed-genociding
; work now?!) to those memeing about me hating zruties.

Looking back at the original commit messages, the main motivation for
this move was to get monsters off punctuation. Which... does not really
hold up as a compelling reason to do a disruptive change like this.
Particularly because there was nothing that I wanted to use the ; symbol
*for*. It was just a largely ungrounded, principled stance of "monsters
should be on letters whenever possible". And the effort was never
completed - it was going to be practically impossible to get things off
@ or &, and even planning to get golems off ' was significantly
difficult.

This doesn't mean I'm planning to revert all of xNetHack's changes to
monster letters and colors - many of them had solid reasons, such as
decluttering overstuffed symbols (elves to Q, mind flayers to U) or
averting terrible interface design (ghosts/shades off space). But I am
rethinking some others. xNetHack is often recommended as a
"vanilla-like" variant, so tossing in a lot of interface disruptions may
not be a good idea.