Move lizards back to : and leprechauns back to l
Same rationale for this as with the sea monster move. The reason for
moving lizards off : in the first place was just "get monsters off
punctuation". If I wanted to add a new object class, then : would be a
great symbol for it, but I don't, and if I want to add a new piece of
furniture I can make do with \.