Revert 3/4 of the Kop recolorings
This is the only recolor reversion I think I have any interest in at the
moment. Currently my priorities are:

1) Avoid ambiguity of multiple species with the same letter and color.
2) Stick with vanilla NetHack coloring.

I am currently trying to figure out where to fit "3) Recolor monsters
to be more thematically colored" into this; if it's the lowest priority,
then I should revert the coloring of things like e.g. acid blobs.

Anyway, "color the Kops according to Yendorian Army ranks" wasn't a
particularly good reason to do it. I have kept the grunt Kops on gray so
as to avoid confusing them with sergeants, but the other three ranks are
back to vanilla's blue/cyan/magenta coloring.