Adjust conduct flags and support xNetHack conducts in conductX
This abandons the practice of tacking new vanilla conducts on to the end
of the conducts array, intermingling them with xNetHack conducts. Now,
conduct flags are made consistent with vanilla, and xNetHack conduct
flags are stored in the highest bits, so conflicts with vanilla flags
will not become an issue until there are more than 32 total conducts.

Bugfix: I noticed in the process that the flags for celibate and
conflictless conduct were inverted. Games were recorded as celibate if
you had NOT been celibate, and so on. Oops.

Also add xNetHack conducts to conductX, which is the planned way for
xlogfile parsers such as the NetHack Scoreboard to read conduct
information in the future.