Base bonesless conduct on loading bones, not the bones option
This is another one of the rare philosophical disagreements with the
devteam (in this case bhaak) that I don't expect to be able to change
their mind on.

To me, if you have not gotten any of the benefits of bones files, then
you have followed bonesless conduct, whether or not you specifically
announced in the config file that you wanted to follow it. For instance,
the first game played on a fresh install isn't going to load any bones;
why deny that player the conduct? They didn't get any benefit from it.

Note that the number of bones files is expressed via bones= in the
xlogfile already; this is just about cases where in vanilla you would
not be credited with bonesless conduct even though bones=0.