Reorganization: pull tin label code into its own function
The list of tin strings isn't long enough to warrant putting it in a
data file (yet), but it was too long and bulky to keep cluttering up
doread().

I also did some improvements to the code itself, with no change in
behavior expected. Output is passed through erode_obj_text now, so if
there are ever rusty/corroded tins, they will obscure the label like
with an eroded T-shirt.