Externify hallublast function instead of array
The extern array was not great since that meant it had to be declared
with a fixed size in multiple source files, which was inflexible to
adding new blasts. Under the new implementation, there's just a function
that retrieves a random string from the array, and that function can be
called from anywhere.