Fix impossible/panic when monster dies to self-read scroll of earth
When a monster read a scroll of earth and died to a boulder falling on
its own head, it would immediately drop the scroll to the ground. Then
when m_useup called extract_from_minvent after that, it would see that
the scroll was now at OBJ_FLOOR instead of OBJ_MINVENT and trigger the
impossible.

The panic happens following this because extract_from_minvent gave up
and did nothing when it encountered an object not at OBJ_MINVENT, and
predictably a lot of callers of extract_from_minvent want to free the
object afterwards. This commit hardens it against that case by calling
obj_extract_self even if the object is not in a monster's inventory.

To fix the problem of the scroll falling to the ground and then having
m_useup called on it, I moved the m_useup call to before the boulders
drop. This is consistent with player use of the scroll anyway,
disappearing before the effect happens.

I also did this for the case where a monster quaffs polymorph; I think
in some rare circumstances it can die indirectly after polymorphing by
falling onto a trap or unfavorable terrain.

(cherry picked from commit c0f232ac975d84f690e7246b17bc02f0657cd5fa)