Fix: saddle being treated as dragon-scaled armor
And potentially a lot of other items could also fall into this category.
The macro that checks for dragon-scaled armor was checking that armcat
(oc_subtyp) equaled ARM_SUIT without actually ensuring it was a piece of
armor. This broke when a horse generated with a saddle (oc_subtyp ==
ARM_SUIT) and its corpsenm was uninitialized.

Fix is to use is_suit, which does check that the item is ARMOR_CLASS.