Prevent obtaining multiple gold stacks by throwing gold at oneself
Discovered with the debug fuzzer and traced the issue back to it
attempting to throw a partial stack of gold at itself. throw_obj
contains a similar bit of logic to re-merge a split stack for when the
prompt is canceled, but it isn't canceled in the "You cannot throw gold
at yourself" case.

Solution is to use that same logic and call unsplitobj() after telling
the player they can't do that.

(cherry picked from commit 8099d6b53c34550d79a7dbe17fa134306eb515e7)