Fix: also prevent multiple gold stacks by quivering a partial stack
This works on basically the same principle as throwing a partial stack
at oneself, but with the ready command. It allowed the player's gold to
be split into two stacks, one in the quiver and one not.

Solution is to check for this case, produce a message informing the
player they can't quiver only part of their gold (quivering all of it
still works fine), and undo the split.