Fix: impossibles when random trees blocked an entire wall
Thus preventing a door from generating next to it. To fix this, I forced
there to always be at least one empty square on each wall when the
relevant terrain type is a 'blocker' - pool, lava, or tree.

These terrain types can still block /traversal/ across a room - so the
limitation of them never spawning in a way where they could block the
Mines entrance is kept.