Fix several thiefstone bugs
The cause of investigation here was an impossible in one of the
thiefstone-teleporting-itself calls (along with a stolen item). The code
did not account for the scenario where the stone was wielded, quivered,
or set as the backup weapon, in which case owornmask would be set and
it would not see itself as a valid thing to teleport. Fix that by adding
a rule that thiefstones see themselves as valid targets before checking
owornmask.

Then, allowing equipped thiefstones to teleport introduced a further bug
where it wasn't clearing uwep/uquiver/uswapwep nor owornmask on the
stone itself. Fixed this by adding a setnotworn to thiefstone_teleport.

Then, this also meant the message for the stone jumping out of your pack
was wrong if you were wielding it, so fixed that. Also fixed other
message oddities e.g. "The scrolls are pulled out of your pack! It
touches the thiefstone..."

Finally, I was noticing something weird in that the thiefstone was
sometimes still jumping out of my pack when I was on its keyed location.
Eventually figured out that it was incorrectly reading the keyed
coordinates of obj (not a thiefstone) rather than otmp.