Fix: prevent panics in abattoirs and zoos
Patch submitted by Michael Meyer, identifying that zoomon() was not
guaranteed to return a valid permonst*, and could thus cause the
abattoir-stocking code to panic when it tried to monsndx() a null
pointer.

Also fixes a potential segfault in zoomon() when rndmonst() returns null
(probably only relevant to extinctionist games).