Fix: intrinsic gain message given for monster that already has it
Fix from EvilHack, which had the same problem. I implemented it somewhat
differently: if the monster has the property extrinsically but not
intrinsically, the message is still printed. The player still gets
intrinsic-gain messages when they have the property extrinsically; I
don't see why it should be different for monsters.

The message is suppressed if the monster already has the intrinsic,
either from its species or from previous meals.