Fix: object materials: sort items correctly according to their material
Sorting the list of items presented at prompts such as looting items out
of a container, or sometimes the inventory display, didn't show them in
any particular order. This was because the sortloot_cmp function was
never updated to account for object materials.

Fix this by stripping the object material in loot_xname() to guarantee
that objects of the same otyp don't get broken up, and then compare the
material strings in the sortloot_cmp "lexicographical" comparison logic
to ensure that items with the same material get put together (unless
overridden by beatitude which takes precedence) and they're shown in
alphabetical order of material.