Allow thiefstones to teleport any gold object
This isn't really a bug, but it was pointed out to me that it's weird
that thiefstones can teleport gold pieces and gold golems but not other
gold items. So, make them capable of that.

Suggestion was that I make the check "if COIN_CLASS || GOLD" to
futureproof against the devteam maybe adding silver/copper pieces, but
on reflection I think that thiefstones like gold, rather than liking
coins, so just checking for gold is fine. One could argue for them to
also teleport items made of other precious metals like silver or
platinum, but nope, they just have a particular affinity for gold.