Fix: "u.uhp reduced while polyselfed" impossible
... when the hero was at XL1 and polyselfed, and got hit with a
life-drain such as a vampire bite, the logic I added to savelife()
called rehumanize() directly, after losexp() had subtracted u.uhp but
before savelife() could re-add it.

I'm not sure why I inserted rehumanize() here in the first place. It
gets called anyway before this, in functions like mdamageu() and
losehp() that ought to cover all sources of damage. After testing that
this works properly if the hero is polyd with unchanging, I basically
reset this back to what vanilla has (plus xNetHack's code for limiting
the amount of HP gained from lifesaving to 100) and removed the
rehumanize call.