Add invocation-related encyclopedia entries
Rescuing them from the defunct better-encyclopedia branch: a vague
description of the invocation and descriptions for the unidentified
forms of the invocation items. The identified forms have more explicit
text about the invocation.

These are somewhat based on the NetHack4/FIQHack entries, but edited
because the originals felt a bit off to me - perhaps the overuse of the
word "special" and general beating around the bush. With the current
structure they're upfront about the existence of an invocation ritual,
whose entry describes exactly what is needed, but only in the formally
identified entries. (For the record, I think vanilla's method of only
cluing an unspoiled player in to the invocation via a single random
oracularity is silly.)