Add additional quotations for many things
Several conducts have quotations now (the ideal is to give all conducts
a quote, but some of them may prove impossible to find something for,
e.g. polypileless):
    atheist
    blind from birth (also triggers for "blind", which lacked one)
    deaf from birth (also triggers for "deaf", which lacked one)
    foodless
    hallucinating from birth (quote also triggers for "hallucination")
    illiterate
    nudist
    pacifist
    vegan/vegetarian
    weaponless

New quotations for:
    alchemy
    athame (from UnNetHack)
    bell
    candle
    invisibility cloak
    credit card
    dented pot
    dwarvish/hooded cloak
    fortune cookie
    Grimtooth (from UnNetHack)
    Moloch (he already had one, but the additional one is quite good)
    scroll of punishment
    wand (already had one, but this is more on-topic)

Also add an entry for "fruit", presumably the user-defined fruit. Not a
literary quotation, but an ambiguous line suggested by jonadab. The C. S.
Lewis quotation about fruit doesn't seem to apply well here so I left it
where it is on orange/pear.