Fix: encyclopedia wasn't updated for Sceptre of Might/Big Stick switch
I put most of the Sceptre's old text under Big Stick, and wrote a mostly
new blurb for the Sceptre, editing as necessary to reflect the
properties they convey now.

Also deleted the energy-boost part of the Mitre of Holiness. I intend to
reintroduce the Mitre at some point as a non-quest artifact, so it's
still searchable for the time being.