Refactor: fully treat strength and energy gain as "intrinsics"
These are in the category of not being actual properties, but also not
being valid for just throwing into the cpostfx() switch statement due to
applying to multiple monster species and having other special handling.
Previously this used -1 and -2 for them as unnamed constants; this names
them and changes the functions around to support them. Specifically, I
wanted them in intrinsic_possible() so that I can later put them in
monster lookup.

The way it works now:
- intrinsic_possible() is the deterministic no-side-effects function
  that determines whether a permonst can convey an intrinsic.
- givit now handles strength and energy gain, not cpostfx.
- There was no reason for cpostfx to call corpse_intrinsic before givit,
  rather than inside it (which mon_givit does), so I moved it inside.
- One very small change was to the energy gain formula; now rather than
  hardcoding newts as d3 gain, it uses 3 as the minimum die size. The
  net effect is that kobold shaman corpses (base level 2) now use a d3
  for energy gain instead of a d2. No other corpses should be affected.