Fix: monster lookup: properly handle werecreatures
Because "werefoo" and "human werefoo" are listed as separate monsters,
monster lookup would only show one of the two, and I think usually the
animal form, which was problematic becasue the animal form is flagged as
G_NOCORPSE and its corpse nutrition value is never used.

Make it so that monster lookup for any werecreature will produce two
lookups, one for each form, and suppress the nocorpse and nutrition
messages on the animal form.

Also don't say their corpse "may" convey lycanthropy; it always does.