Fix: "thiefstone" object lookup tried to match to "thief"
More parsing-string-as-rank shenanigans. My fix was to extract the
wish parser code that checks for strings that look like they might fuzzy
match a monster but actually are definitely an object into its own
function, and then call that before trying to match the string to a
monster.

This also fixes other similar bugs such as matching "samurai" when
trying to look up "samurai sword", etc.