Cut down on pacifist warning spam a bit
Annoyance: if you are not playing pacifist, and go 100 turns without
hitting anything, and attack a monster, and get prompted to confirm you
want to attack it, and miss (or hit it non-fatally), you then get
prompted again. And again, and again, until you actually manage to kill
it.

I added a variable that tracks whether the player has confirmed they
want to attack the monster. It disables further prompting. I implemented
it as a static boolean because that allows it to persist for the rest of
the game session but won't be saved and restored (so someone returning
to a saved game after months of not playing it will be reprompted if
they forget they had been pacifist.)