Don't translate a dragon scale mail wish into scales
Feedback from the beta version on Hardfought is that this has already
tripped up some players who are then unhappy that they got an inferior
item to what they want. Be nicer by removing the DSM to scales mappings,
so that a wish for dragon scale mail produces the doesn't-exist message,
and a player that wishes for one assuming it still exists will then be
prompted to go find out why it doesn't and learn about the new system.

Also remove some now-invalid DSM wishes from the wish parser tests.