Do a ton of sticking an "x" in front of "NetHack" strings
This is really basic first-release-of-your-variant stuff, but largely
hasn't been done up until now except where necessary for the Linux
build.

It touches non-Windows builds here, but the main reason for doing this
is so that xNetHack doesn't try to write into "NetHack" folders on
Windows and potentially trample over someone's existing NetHack
installation.