Default the 'color' option to on
It's 2021. Does anyone start a game of NetHack, see the colorless
interface, and think "Ahh... this is just how I want it?" I doubt it.

The option's still there, so it can be turned off if the player really
doesn't want color, but now the onus is on that tiny (nonexistent?)
fraction of players to go and change their config file, and the large
majority of new players who want color don't have to go figure out the
config file right off (or at least have one less thing to do there).