Fix: no defined behavior for hangup during a wish prompt
In xNetHack, repeated escaping out of a wish prompt will repeatedly
prompt if you really want to forfeit this wish, and requires the user to
press 'y'. spicycebolla pointed out that this behavior doesn't account
for what happens in the event of a hangup. A wish prompt is more likely
than most places to end in a hangup (pause the game, go ask people or
read a wiki and figure out what you want to wish for, and the game hangs
up in the meantime...)

So this adds a done_hup check to force a wish for "nothing" so that the
program can continue and hopefully exit gracefully. Suggestion was to
provide a default wish such as a magic marker or amulet of life saving,
which I am not doing; apart from that being too nice, it could break
someone's wishless conduct.

In testing, I got two different flaky outcomes: one in which the program
gracefully saves and exits, and one in which a child process that's
forked in docompress_file() starts to run away in an infinite loop of
sending Esc to the program and a save is not made. I haven't been able
to reliably reproduce that, but in case it recurs, I also reintroduced a
maximum amount of cancellations out of the wish prompt (50) before it
gives up and sends in a blank wish. Not reintroducing the "That's enough
tries!" logic though - this is only a failsafe if that runaway process
occurs again.