Add automatic builds and deployment with GitHub Actions
GitHub appears to be moving away from using travis-ci to do continuous
integration, with the new thing being Actions, so I figured it was
better to do it the new way.

I have two motivations here: one, to automatically verify whether the
master branch (or pull requests to it) are broken, and two, to produce
Windows binaries of xNetHack, as that is a feature I'm getting an
increasing amount of requests for.

This creates a yml file that will trigger linux, Mac, and Windows
builds when a commit is pushed to master or a PR is opened into master.
Additionally, when a tag is pushed, it will build Windows binaries and
upload them as a release.