Fix: whistle naming inconsistency
When tin whistles were changed to pea whistles, I removed the
unidentified description of "whistle", but left it for magic whistles.
This created weirdness in which the hero would find "pea whistles" and
"whistles" lying around on the ground. Nothing in my original commit
specifically indicated why I was removing the description (it may have
been because I felt at the time that whistles are trivially identified
and thus shouldn't be ambiguous). But my preference now is to not have
magic whistles be immediately apparent as such, so this restores the pea
whistle's unidentified description.