Fix: trying to give intrinsic 0 to the player
Caused by the recent refactor of givit() and associated functions.
Now that the corpse_intrinsic call moved inside givit, "property" 0 -
not actually a property - would pass through to should_givit, and hit
its default case in which 1/15 of corpses/tins would confer it. This
triggered an impossible because 0 isn't actually anything.

Guard against this in givit and mon_givit by returning early if the
corpse is not going to give an intrinsic.