Fix: scroll of water produced no message if it had no effect
In general, scrolls do not do this (even taming has a message if there
is nothing to tame). Fix it by adding messages which are unambiguous but
don't automatically identify the scroll (since unambiguous messages for
other scrolls don't at this time); the player will still be prompted to
call the scroll afterwards.

Messages are from suggestions by qt.