Fix: rubbing touch/thiefstone on hands did not work
In the getobj patch I got merged into vanilla, touchstone_ok excludes
rubbing hands on the stone, and it uses any_obj_ok (also excluding
hands) if the stone is unidentified. However, in xNetHack, there are
special cases for rubbing both touchstones and thiefstones on one's
hands, and a general case for nothing interesting happening.

To fix this, I made rubbing hands downplayed instead of excluded (even
an identified blessed thiefstone shouldn't outright suggest rubbing it
on your hand if you're a gold golem) in both the *stone_ok() callbacks,
and had to define a new callback for "suggest everything including
hands" which now covers the general case.