Fix: several more thiefstone bugs
Initial bug report from qt was that thiefstones found in bones files
would keep their same ledger number from the previous game - referencing
a level that might not exist in the current game, and even if it did
exist the x and y wouldn't necessarily correspond to a valid position.

I tackled this by adding rules in the bones saving/loading code to make
it so that non-cancelled thiefstones in bones will always be keyed
somewhere on the bones level, regardless of whether they were keyed
there in the previous game. If they were keyed to that level already,
their x and y will be preserved; otherwise it will pick a new location.
In order to do that, I extracted the thiefstone-location-picking
algoritm into its own function because init_thiefstone() has other side
effects on the stone.

I beefed up the guards on thiefstone ledger validity by defining a
constant for its spe when cancelled, and a macro that checks whether the
ledger number is a valid ledger. In the process of this, I discovered
and fixed several other bugs:

- The game panicked when trying to teleport a gold golem with a
  cancelled thiefstone (at least in the hit-own-head case, I didn't test
  all the cases) because it tried to send you to ledger -1. Now it just
  does nothing.
- Inconsistent use of 0 and -1 for a cancelled thiefstone. Now it should
  always be -1.
- Inconsistent use of > 0, == 0, < 1, etc for thiefstone ledger validity
  checks. Now it should always use thiefstone_ledger_valid().
- When you wished (or possibly polypiled) a thiefstone, it called
  init_thiefstone with all of its side effects - including generating a
  chest in the keyed location it picked. This shouldn't happen; the
  chest should only appear with random thiefstones, so I guarded it with
  g.in_mklev.