Fix: water_damage_chain could wet some objects twice
Crash reported in EvilHack which has copied this code; it's far more
likely there for water_damage_chain to destroy an item outright (because
it allows iron items to rust away completely), and attempting to wet an
object that has already been deleted caused the crash.

qt discovered the root cause, which is a flaw in the reservoir sampling
code I wrote; when it starts to copy different items into the to_damage
array, it starts from the beginning of the list. But the items at the
beginning of the list have already been used to initialize to_damage. So
this allowed them to potentially be put into to_damage in multiple
cases.

Still a bug in xNetHack because water damage should never hit an item
twice. To fix, I made the second loop (the one that writes new items
into to_damage) skip the first items of the object chain until it has
reached objects that weren't put into to_damage by default.