Shopkeepers address certain players with titles
The most common of these is "Sir" or "Dame" for all knights, but I also
implemented two for high-level monks and healers. Shopkeepers only use
the title when they are friendly towards the player (entering their shop
in good standing or encountering them outside the shop when they're
chasing a different player in a bones file).

This feature was previously considered, then rejected, and now it's
back. The reason I didn't like it at first was because it looked (and
still looks) a bit weird either way for capitalizing or not capitalizing
the player's name if it starts with a lowercase letter. E.g. "Sir Stth"
and "Sir stth" both look weird. Ultimately, though, I decided that
capitalizing the name is no stranger than its appearance in combat
messages ("Stth's ghost touches you!") so this is fine to have.