Don't guard against escaping out of wishes in wizard mode
The guard exists to prevent accidental loss of a wish to a typo, but in
wizard mode you can just wish again, so it doesn't really do anything. I
think this will also speed up the fuzzer since it won't waste a lot of
input on entering random characters at the yn("Really forfeit this
wish?") prompt.