Ensure getroomtype returns valid result in mklev
getroomtype(sounds.c) uses orig_rtype, which is not initialized until
the very end of the level generation process.  When used during level
generation, it returned a value of 0 (OROOM) for every room.  This meant
that cockatrice generation in a cockatrice nest produced random monster
statues, even though they are only supposed to be generated when a
cockatrice is in a normal room or hallway.