Fix: Arc starting pick was enchanted
Starting pick in vanilla is always +0 (though, this appears to be due to
there not being any code in mksobj() that sets its enchantment to
anything like there is for weapons). But in xNetHack, it was appearing
as rather well enchanted - always +1, sometimes +3 even and possibly
higher.

This is because I rearranged the u_init code at one point. One thing
that moved out of the "unspecified type" codepath was a conditional
setting any oc_charged object's enchantment to rne(3) if it was 0 or
below. The intention of this statement was to catch rings, and I moved
it to the unified code path so that if there were ever explicitly
specified ring types with unspecified enchantment, they'd always end up
positive as well. But since a whole bunch of other objects are also
oc_charged (including pick-axes), the conditional caught them as well.
In virtually all other cases, though, there was a specific spe set in
the hardcoded list that then discarded the rne(3) value. The starting
pick didn't discard this because it isn't actually hardcoded as +0 - it
just always ends up as such, as mentioned above.

This is possibly a latent vanilla bug, which doesn't happen only because
no character receives e.g. a random tool which could be a pick-axe in
their initial inventory.

Fix this by making the condition explicitly only apply to rings (as the
vanilla comment and my commit message when I moved the conditional
indicate the original intention is.)