Fix: door states in lua getmap() didn't work correctly
The parser code did not account for the xNetHack door overhaul, so it
was still doing things like (levl[x][y].flags & D_ISOPEN) to test if a
door was open. This was latent (because nothing in the game used getmap)
until I started to use it.

Also support iron doors in getmap()'s return, and return the door flags
for secret doors as well (since secret doors can be specified as locked,
trapped, iron, etc).