Fix: specifying a secret door with des.door was actually impossible
Problem seems to have been introduced in the xNetHack lspo_door refactor
in 6ba3d6c, in which it generates a door mask and passes that to
sel_set_door, but completely ignores the variable 'typ' it sets up to
handle secret doors. In reality, sel_set_door is still expecting the
D_SECRET flag to be set on the door mask passed in, so set that.