Support specifying object materials in des.object()
Only in the table form for now. This was remarkably simple and I suppose
the only reason I didn't do it already is because I never needed
specific materials on things.

Left unspecified, the object is still eligible for materials as normal,
and as noted in a comment, there is not currently a way to specify that
you want an item of undetermined type but you want its material to be
always the default.