Fix: only print bones level sound when bones are undiscovered
I was noticing that the "eerie feeling" message persists long after the
player has found the bones and dealt with any ghost or other monster
left behind by the player, which kind of kills the flavor of it. So this
limits the sound to when there are bones on the level that have still
not been found.

The original plan was to print the message if a struct ebones-having
monster existed on the level, but bones can be created without leaving
behind a monster.

As part of this, I discovered that the bonesknown flag was actually
never set in the g.level.bonesinfo chain -- only in mapseen's cemetery
chain. Since it doesn't look like mapseen is supposed to be used outside
of dungeon.c, and I couldn't find any reason for why bonesknown should
never be set in g.level.bonesinfo, I made it so that the flag is set
there as well.