Prevent noise from uncloaking mimics
In order to warn the player consistently whenever a sleeping monster is
woken up, xNetHack started using wakeup(mon.c) for all monster waking
(in contrast to vanilla, which just sets mon->msleeping to 0 in a number
of places).  However, the wakeup function also uncloaks mimics, so after
this update, any circumstances that would have previously woken sleeping
monsters (kicking, fighting, blowing a whistle...) now also revealed
mimics in the vicinity.  This made fighting a monster in a shop a lot
riskier, for example, since any number of mimics might abruptly uncloak
and make a beeline towards the hero as soon as the first punch was
thrown.

This commit adds a reveal_hidden parameter to wakeup, which can be used
to specify whether a call to wakeup should reveal a mimic or not, and
uses it to restrict situations which wake mimics to something closer to
the previous norm.

Many of these calls are not currently reachable for mimics (e.g. trying
to haul a mimic out of a pit will dump you out early with "wait! that's
a small mimic" before reaching the wakeup call, similar things are true
for other cases) but it makes sense for the logic to be sound and as
internally consistent as possible in case that changes in the future --
these actions logically should reveal mimics, even if it's not currently
possible to reach that point.

Also add a guard against calling wakeup on a dead monster to prevent
messages like "the newt is killed! the newt wakes up", and replace
another use of 'msleeping = 0' (monster hitting another monster with an
egg) that seemingly got missed previously with wakeup.