Prevent noise from uncloaking mimics
In order to warn the player consistently whenever a sleeping monster is
woken up, xNetHack started using wakeup(mon.c) for all monster waking
(in contrast to vanilla, which just sets mon->msleeping to 0 in a number
of places).  However, the wakeup function also uncloaks mimics, so after
this update, any circumstances that would have previously woken sleeping
monsters (kicking, fighting, blowing a whistle...) now also revealed
mimics in the vicinity.  This made fighting a monster in a shop a lot
riskier, for example, since any number of mimics might abruptly uncloak
and make a beeline towards the hero as soon as the first punch was
thrown.

This commit adds a reveal_hidden parameter to wakeup, which can be used
to specify whether a call to wakeup should reveal a mimic or not, and
uses it to restrict situations which wake mimics to something closer to
the previous norm.