Fix: water bucket trap hit triggering monster even at range
doortrapped() had a comment proclaiming monsters can't affect doors at
range and therefore if it's triggering a door trap it must be touching
the door. This is not the case; monsters can zap wands that affect the
door. In my case, a dwarf zapped striking at me, and hit a door behind
me... causing a bucket of water to somehow splash down on the dwarf.

Fix this by removing that assumption (doortrapped called with NO_PART
always implies the interacter is not touching it) and fixing the water
bucket case to account for that. The other cases that involved touching
the door appear to already handle this smoothly.