Fix: worm teeth generating eroded or erodeproof
Erosion and erosionproofing and greasing were potentially present on
every weapon created ever, including the tooth fresh out of a worm's
mouth. The part of this commit that fixes that is adding a g.in_mklev
check so that it only affects items generated at level creation time.
I'm not totally happy with this, because it means that monsters spawning
after level creation will never generate with erodeproof or eroded gear,
nor will death drops be eligible for that. But it seems better than the
alternative of manually tracking down and resetting these values in
cases where it really shouldn't apply.

Part of that is because I expanded the range of possible items that
these flags can get set on in this commit also: previously it was just
weapons and I'm not sure why; there was no explanation for it so I
assume it was an error. Now all items (that can get eroded) are eligible
for erosionproofing, erosion, and grease. In my first test game after
moving the code, I spawned with a burnt unicorn horn in my wizkit; the
in_mklev check will also prevent having to insert code in u_init.c to
compensate for that sort of thing.

Another fix in this commit was that I noticed the two types of erosion
always came hand in hand: an item that was capable of two forms of
erosion would have both, or neither, due to a logic error when
implementing it - the two cases didn't roll separately. Now they're
handled independently.