Improve internal logic of some wakeup calls
Many of these are not currently reachable for mimics (e.g. trying to
haul a mimic out of a pit will dump you out early with "wait! that's a
small mimic" before reaching the wakeup call, similar things are true
for other cases) but it makes sense for the logic to be sound and as
internally consistent as possible in case that changes in the future --
these actions logically should reveal mimics, even if it's not currently
possible to reach that point.

Also add a guard against calling wakeup on a dead monster to prevent
messages like "the newt is killed! the newt wakes up", and replace
another use of 'msleeping = 0' (monster hitting another monster with an
egg) that seemingly got missed previously with wakeup.